requirements.txt
anyio==3.6.2
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
cryptography==40.0.1
defusedxml==0.7.1
Django==4.2
django-allauth==0.54.0
django-cleanup==7.0.0
django-crispy-forms==2.0
django-environ==0.10.0
django-filter==23.1
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
oauthlib==3.2.2
phonenumbers==8.13.8
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
python-telegram-bot==20.2
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.15
